requirements.txt
Adafruit-DHT
bmp-280
RPi.GPIO
fastapi
pydantic